+++ libs/rxjs/number/src/index.ts
@@ -6,6 +6,7 @@
 export { filterIsFinite } from './lib/filter-is-finite';
 export { filterIsInteger } from './lib/filter-is-integer';
 export { filterIsSafeInteger } from './lib/filter-is-safe-integer';
+export { fromNumber } from './lib/from-number';
 export { isFinite } from './lib/is-finite';
 export { isInteger } from './lib/is-integer';
 export { isNaN } from './lib/is-nan';