+++ libs/rxjs/number/src/lib/is-nan.ts
@@ -20,7 +20,7 @@ import { map } from 'rxjs/operators';
  * ```
  *
  * @returns Boolean value of a number passing the `Number.isNaN` equality check
- * @category RxJS Number Equality
+ * @category RxJS Number Query
  */
 export function isNaN(): OperatorFunction<number, boolean> {
   return (source: Observable<number>) => source.pipe(map((value) => Number.isNaN(value)));