+++ libs/rxjs/number/src/lib/is-safe-integer.ts
@@ -22,7 +22,7 @@ import { map } from 'rxjs/operators';
  * ```
  *
  * @returns Boolean value of a number passing the `Number.isSafeInteger` equality check
- * @category RxJS Number Equality
+ * @category RxJS Number Query
  */
 export function isSafeInteger(): OperatorFunction<number, boolean> {
   return (source: Observable<number>) => source.pipe(map((value) => Number.isSafeInteger(value)));