+++ libs/rxjs/string/src/lib/map-char-code.spec.ts
@@ -1,11 +1,11 @@
 import { of } from 'rxjs';
 import { take } from 'rxjs/operators';
-import { filterCharCode } from './filter-char-code';
+import { mapCharCode } from './map-char-code';
 
-describe('filterCharCode', () => {
+describe('mapCharCode', () => {
   it('should create a value from an array of values', (done) => {
     of([35, 36, 37, 38])
-      .pipe(filterCharCode(), take(1))
+      .pipe(mapCharCode(), take(1))
       .subscribe({
         next: (value) => expect(value).toBe('#$%&'),
         complete: () => done(),