+++ libs/rxjs/string/src/lib/map-code-point.spec.ts
@@ -1,11 +1,11 @@
 import { of } from 'rxjs';
 import { take } from 'rxjs/operators';
-import { filterCodePoint } from './filter-code-point';
+import { mapCodePoint } from './map-code-point';
 
-describe('filterCodePoint', () => {
+describe('mapCodePoint', () => {
   it('should create a value from an array of values', (done) => {
     of([9731, 9733, 9842])
-      .pipe(filterCodePoint(), take(1))
+      .pipe(mapCodePoint(), take(1))
       .subscribe({
         next: (value) => expect(value).toBe('☃★♲'),
         complete: () => done(),