+++ libs/rxjs/string/src/lib/map-code-point.ts
@@ -6,7 +6,7 @@ import { Observable, OperatorFunction } from 'rxjs';
 import { map } from 'rxjs/operators';
 
 /**
- * The `filterCodePoint` operator can be used with an {@link https://rxjs-dev.firebaseapp.com/guide/observable|Observable} array of
+ * The `mapCodePoint` operator can be used with an {@link https://rxjs-dev.firebaseapp.com/guide/observable|Observable} array of
  * numbers that represent character code points, and returns a string value
  *
  * @remarks
@@ -21,13 +21,13 @@ import { map } from 'rxjs/operators';
  *      acc.push(val);
  *      return acc;
  *    }, []),
- *    filterCodePoint()
+ *    mapCodePoint()
  *  ).subscribe(...) // '☃★♲'
  * ```
  *
  * @returns String from an array of character codes
- * @category RxJS From String Creation
+ * @category RxJS String Map
  */
-export function filterCodePoint(): OperatorFunction<number[], string> {
+export function mapCodePoint(): OperatorFunction<number[], string> {
   return (source: Observable<number[]>) => source.pipe(map((values) => String.fromCodePoint(...values)));
 }